EjbBundleDescriptor - moving more fields to the parent
- This is safe as there is only one child of this abstract class
