+++ appserver/deployment/dol/src/main/java/com/sun/enterprise/deployment/LocalStrings.properties
@@ -115,7 +115,6 @@ enterprise.deployment.exceptionapphasnoresourceenvrefbyname=This app [{0}] has n
 enterprise.deployment.exceptionapphasnoentitymgrfactoryrefbyname=This app [{0}] has no entity manager factory reference by the name of [{1}]
 enterprise.deployment.exceptionapphasnoentitymgrrefbyname=This app [{0}] has no entity manager reference by the name of [{1}]
 enterprise.deployment.exceptionapphasnoenvpropertybyname=This app [{0}] has no environment property by the name of [{1}]
-enterprise.deployment.exceptionapphasnoejbrefbyname=This app [{0}] has no ejb reference by the name of [{1}]
 enterprise.deployment.exceptionapphasnoresourcerefbyname=This app [{0}] has no resource reference by the name of [{1}]
 enterprise.deployment.exceptionapplicationduplicatedatasourcedefinition=This app [{0}] cannot have datasource definitions of same name : [{1}]
 enterprise.deployment.exceptionapplicationduplicateconnectionfactorydefinition=This app [{0}] cannot have connection factory definitions of same name : [{1}]
