+++ appserver/ejb/ejb-container/src/main/resources/org/glassfish/ejb/deployment/descriptor/LocalStrings.properties
@@ -33,7 +33,6 @@ enterprise.deployment.exceptionmsgbeantxtypenotlegaltype=[{0}] is not a legal tr
 enterprise.deployment.invalidcmpversion=Invalid CMP version: {0}.
 enterprise.deployment.pkhasnopersistentfields=CMP bean [{0}], primary key class [{1}] has public non-persistent field(s) [{2}].
 enterprise.deployment.exceptioncannotgetsecbusmethodsinmsgbean=Cannot get business method for security for message-driven bean.
-enterprise.deployment.exceptionapphasnoejbrefbyname=This app [{0}] has no ejb reference by the name of [{1}]
 enterprise.deployment.exceptionejbhasnoservicerefbyname=This ejb [{0}] has no service reference by the name of [{1}]
 enterprise.deployment.exceptionbeanhasnoresourceenvrefbyname=This bean [{0}] has no resource environment reference by the name of [{1}]
 enterprise.deployment.exceptionbeanhasnoentitymgrfactoryrefbyname=This ejb [{0}] has no entity manager factory reference by the name of [{1}]
